# docs/swagger.yaml
definitions:
  models.APIResponse:
    properties:
      data:
        description: Optional data payload
      error:
        description: |-
          Optional error
          example: Something went wrong
        type: string
      message:
        description: |-
          The response message
          example: Success
        type: string
    type: object
...
  models.SoftwareDepartmentMatch:
    properties:
      created_at:
        type: string
      department_id:
        description: |-
          DepartmentID is the ID of the department receiving the software
          example: 3
        type: integer
      id:
        description: |-
          ID is the unique identifier of the department match
          example: 1
        type: integer
      software_id:
        description: |-
          SoftwareID is the ID of the software
          example: 2
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/software-department-matches:
    get:
      description: Retrieves all department-level software matches, optionally filtered
        by software_id or department_id
      parameters:
      - description: Filter by Software ID
        in: query
        name: software_id
        type: integer
      - description: Filter by Department ID
        in: query
        name: department_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SoftwareDepartmentMatch'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get all software department matches
      tags:
      - Software Department Matches
    post:
      consumes:
      - application/json
      description: Creates a new department-level software match
      parameters:
      - description: Software Department Match
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/models.SoftwareDepartmentMatch'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SoftwareDepartmentMatch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create a software department match
      tags:
      - Software Department Matches
  /api/software-department-matches/{id}:
    delete:
      description: Deletes a department-level software match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No Content - Deletion Success
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete a software department match
      tags:
      - Software Department Matches
    get:
      description: Retrieves a single software department match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SoftwareDepartmentMatch'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get a software department match by ID
      tags:
      - Software Department Matches
    put:
      consumes:
      - application/json
      description: Updates an existing software department match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Software Department Match
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/models.SoftwareDepartmentMatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SoftwareDepartmentMatch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Update a software department match
      tags:
      - Software Department Matches
  /api/software-department-matches/{id}/revoke:
    delete:
      description: Deletes a match and revokes software from all staff in the associated
        department
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No Content - Deletion Success
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Delete match and revoke software from department staff
      tags:
      - Software Department Matches
  /api/software-department-matches/auto-assign:
    post:
      consumes:
      - application/json
      description: Creates a new match and automatically assigns the software to all
        staff in the department
      parameters:
      - description: Software Department Match
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/models.SoftwareDepartmentMatch'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SoftwareDepartmentMatch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Create department match with auto-assignment
      tags:
      - Software Department Matches
...
swagger: "2.0"